+++ integration/admin-client/src/main/java/org/keycloak/admin/client/resource/RealmResource.java
@@ -297,6 +297,9 @@ public interface RealmResource {
     @Path("saml-federations")
     SAMLFederationResource samlFederation();
 
+    @Path("openid-federations")
+    public OpenIdFederationsResource openIdFederationsResource();
+
     @DELETE
     void remove();