+++ model/infinispan/src/main/java/org/keycloak/models/cache/infinispan/entities/CachedRealm.java
@@ -52,6 +52,7 @@ import org.keycloak.models.RealmModel;
 import org.keycloak.models.RequiredActionProviderModel;
 import org.keycloak.models.RequiredCredentialModel;
 import org.keycloak.models.WebAuthnPolicy;
+import org.keycloak.models.OpenIdFederationGeneralConfig;
 import org.keycloak.models.cache.infinispan.DefaultLazyLoader;
 import org.keycloak.models.cache.infinispan.LazyLoader;
 import org.keycloak.representations.idm.RealmRepresentation;
@@ -121,6 +122,7 @@ public class CachedRealm extends AbstractExtendableRevisioned {
     protected OTPPolicy otpPolicy;
     protected WebAuthnPolicy webAuthnPolicy;
     protected WebAuthnPolicy webAuthnPasswordlessPolicy;
+    protected OpenIdFederationGeneralConfig openIdFederationConfig;
 
     protected String loginTheme;
     protected String accountTheme;
@@ -242,6 +244,7 @@ public class CachedRealm extends AbstractExtendableRevisioned {
         otpPolicy = model.getOTPPolicy();
         webAuthnPolicy = model.getWebAuthnPolicy();
         webAuthnPasswordlessPolicy = model.getWebAuthnPolicyPasswordless();
+        openIdFederationConfig = model.getOpenIdFederationGeneralConfig();
 
         loginTheme = model.getLoginTheme();
         accountTheme = model.getAccountTheme();
@@ -681,6 +684,10 @@ public class CachedRealm extends AbstractExtendableRevisioned {
 
     public WebAuthnPolicy getWebAuthnPasswordlessPolicy() {
         return webAuthnPasswordlessPolicy;
+    }
+
+    public OpenIdFederationGeneralConfig getOpenIdFederationConfig() {
+        return openIdFederationConfig;
     }
 
     public AuthenticationFlowModel getBrowserFlow() {