+++ server-spi-private/src/main/java/org/keycloak/events/Errors.java
@@ -66,6 +66,10 @@ public interface Errors {
     String INVALID_SIGNATURE = "invalid_signature";
     String INVALID_REGISTRATION = "invalid_registration";
     String INVALID_ISSUER = "invalid_issuer";
+    String INVALID_SUBJECT = "invalid_subject";
+    String INVALID_TRUST_ANCHOR = "invalid_trust_anchor";
+    String INVALID_METADATA = "invalid_metadata";
+    String INVALID_TRUST_CHAIN = "invalid_trust_chain";
     String INVALID_FORM = "invalid_form";
     String INVALID_CONFIG = "invalid_config";
     String EXPIRED_CODE = "expired_code";