+++ server-spi-private/src/main/java/org/keycloak/events/admin/ResourceType.java
@@ -215,5 +215,10 @@ public enum ResourceType {
     /**
      *
      */
-    , ORGANIZATION_MEMBERSHIP;
+    , ORGANIZATION_MEMBERSHIP
+
+    /**
+     *
+     */
+    , OPENID_FEDERATION;
 }