+++ server-spi/src/main/java/org/keycloak/models/RealmModel.java
@@ -17,6 +17,7 @@
 
 package org.keycloak.models;
 
+import java.util.ArrayList;
 import java.util.Comparator;
 import org.keycloak.common.enums.SslRequired;
 import org.keycloak.component.ComponentModel;
@@ -333,6 +334,30 @@ public interface RealmModel extends RoleContainerModel {
      * @param policy
      */
     void setWebAuthnPolicyPasswordless(WebAuthnPolicy policy);
+
+    default OpenIdFederationGeneralConfig getOpenIdFederationGeneralConfig() {
+        return null;
+    };
+
+    default boolean isOpenIdFederationEnabled() {
+        return getOpenIdFederationGeneralConfig() != null;
+    };
+
+    default void setOpenIdFederationGeneralConfig(OpenIdFederationGeneralConfig generalConfig) {
+    };
+
+    default List<OpenIdFederationConfig> getOpenIdFederations() {
+        return new ArrayList<>();
+    }
+
+    default void addOpenIdFederation(OpenIdFederationConfig fedConfig) {
+    }
+
+    default void updateOpenIdFederation(OpenIdFederationConfig fedConfig) {
+    }
+
+    default void removeOpenIdFederation(String internalId) {
+    }
 
     RoleModel getRoleById(String id);