+++ services/src/main/java/org/keycloak/jose/jws/DefaultTokenManager.java
@@ -53,6 +53,7 @@ import org.keycloak.protocol.oidc.OIDCLoginProtocol;
 import org.keycloak.protocol.oidc.mappers.LogoutTokenMapper;
 import org.keycloak.representations.LogoutToken;
 import org.keycloak.services.util.DefaultClientSessionContext;
+import org.keycloak.representations.openid_federation.EntityStatement;
 import org.keycloak.util.JsonSerialization;
 import org.keycloak.util.TokenUtil;
 
@@ -84,6 +85,16 @@ public class DefaultTokenManager implements TokenManager {
 
         String type = type(token.getCategory());
         String encodedToken = new JWSBuilder().type(type).jsonContent(token).sign(signer);
+        return encodedToken;
+    }
+
+    @Override
+    public String encodeForOpenIdFederation(EntityStatement token) {
+        String signatureAlgorithm = signatureAlgorithm(token.getCategory());
+
+        SignatureSignerContext signer = session.getProvider(SignatureProvider.class, signatureAlgorithm).signer();
+
+        String encodedToken = new JWSBuilder().type(token.getType()).jsonContent(token).sign(signer);
         return encodedToken;
     }
 
@@ -201,6 +212,7 @@ public class DefaultTokenManager implements TokenManager {
             case ADMIN:
                 return getSignatureAlgorithm(null);
             case ACCESS:
+            case ENTITY_STATEMENT:
                 return getSignatureAlgorithm(OIDCConfigAttributes.ACCESS_TOKEN_SIGNED_RESPONSE_ALG);
             case ID:
             case LOGOUT: