+++ services/src/main/java/org/keycloak/protocol/oidc/federation/TrustChainProcessorFactory.java
@@ -1,0 +1,6 @@
+package org.keycloak.protocol.oidc.federation;
+
+import org.keycloak.provider.ProviderFactory;
+
+public interface TrustChainProcessorFactory extends ProviderFactory<TrustChainProcessor> {
+}