+++ services/src/main/java/org/keycloak/services/resources/RealmsResource.java
@@ -29,14 +29,18 @@ import org.keycloak.models.KeycloakSession;
 import org.keycloak.models.RealmModel;
 import org.keycloak.protocol.LoginProtocol;
 import org.keycloak.protocol.LoginProtocolFactory;
+import org.keycloak.protocol.oidc.federation.OpenIdFederationWellKnownProvider;
 import org.keycloak.services.CorsErrorResponseException;
+import org.keycloak.services.clientregistration.ClientRegistrationAuth;
 import org.keycloak.services.clientregistration.ClientRegistrationService;
+import org.keycloak.services.clientregistration.openid_federation.OpenIdFederationClientRegistrationService;
 import org.keycloak.services.cors.Cors;
 import org.keycloak.services.managers.RealmManager;
 import org.keycloak.services.resource.RealmResourceProvider;
 import org.keycloak.services.resources.account.AccountLoader;
 import org.keycloak.services.util.CacheControlUtil;
 import org.keycloak.services.util.ResolveRelative;
+import org.keycloak.util.TokenUtil;
 import org.keycloak.utils.ProfileHelper;
 import org.keycloak.wellknown.WellKnownProvider;
 import org.keycloak.wellknown.WellKnownProviderFactory;
@@ -173,6 +177,16 @@ public class RealmsResource {
         return new ClientRegistrationService(session, event);
     }
 
+    @Path("{realm}/openid-federation/clients-registrations")
+    public OpenIdFederationClientRegistrationService getOpenIdFederationClientsService(final @PathParam("realm") String name) {
+        resolveRealmAndUpdateSession(name);
+        EventBuilder event = new EventBuilder(session.getContext().getRealm(), session, session.getContext().getConnection());
+        OpenIdFederationClientRegistrationService provider = new OpenIdFederationClientRegistrationService(session);
+        provider.setEvent(event);
+        provider.setAuth(new ClientRegistrationAuth(session, provider, event, "openid-connect"));
+        return provider;
+    }
+
     @Path("{realm}/clients-managements")
     public ClientsManagementService getClientsManagementService(final @PathParam("realm") String name) {
         resolveRealmAndUpdateSession(name);
@@ -253,6 +267,10 @@ public class RealmsResource {
 
             // Check if the provider returned a JWT string or JSON object
             responseBuilder = Response.ok(config).type(config instanceof String ? APPLICATION_JWT : MediaType.APPLICATION_JSON);
+
+            if (wellKnown instanceof OpenIdFederationWellKnownProvider) {
+                responseBuilder.header("Content-Type", TokenUtil.APPLICATION_ENTITY_STATEMENT_JWT);
+            }
 
             return Cors.builder().allowAllOrigins().auth().add(responseBuilder.cacheControl(CacheControlUtil.noCache()));
         }