+++ services/src/main/java/org/keycloak/services/resources/admin/RealmAdminResource.java
@@ -1196,6 +1196,11 @@ public class RealmAdminResource {
     public SAMLFederationResource getIdentityProviderFederationResource() {
         return new SAMLFederationResource(realm, session, this.auth, adminEvent);
     }
+
+    @Path("openid-federations")
+    public OpenIdFederationsResource getOpenIdFederationsResource() {
+        return new OpenIdFederationsResource(realm, session, this.auth, adminEvent);
+    }
     
     
     /**